test(cli): split sandbox connect inference tests
