+++ ci/test-file-size-budget.json
@@ -12,7 +12,6 @@
     "test/onboard-messaging.test.ts": 2122,
     "test/onboard-selection.test.ts": 7757,
     "test/onboard.test.ts": 4887,
-    "test/policies.test.ts": 3145,
-    "test/sandbox-connect-inference.test.ts": 1577
+    "test/policies.test.ts": 3143
   }
 }
